Modify cgit-gen script to handle repository grouping
CGit allows you to group repositories with the 'repo.group' config file option.
Once set this option is applied to all the repository definitions following it
until you set it to something else.

The script will look for a 'group' file in the .git dir, along the same lines
as the 'description' file. This file contains the name of the group that the
repo should be shown in. If this file doesn't exist or is empty then the repo
is shown at the start of the list without a group.