Update the list of clone-prefixes to match reality.
We haven't supported general ssh for over a year, and while ssh is
available via gerrit, it doesn't really add much to list it here
as users need an SR account to use that.
Instead, just list the public gerrit https:// and the git:// urls,
which are current.

Note that while non-gerrit https:// is still available, it won't
be much longer (and isn't actually serving the right things, see
ticket #2060).

Change-Id: I9c3e132ad66f5f885098d92cdd08bf7c30e7f4a8