Enable snapshot downloading
With this setting it's possible to get a .tar.gz/.tar.bz2/.zip archive
from commits rather than needing to clone over git and produce the
archive with `git archive`.

Beyond convenience, there are a couple of useful applications for this
in CI which I'd like to investigate (using archives rather than doing a
full git clone).